Replicate layout: sort replicated sheets.
Sort sheets to be replicated by the number portion of the reference of
the pivot module. This means the order is stable (doesn't depend on the
order the modules are loaded). It also looks better on the PCB as the
references increment nicely between blocks rather than jumping all over
the place.